Separate type inference algorithm from iterator IR type inference impl
